backend/requirements.txt
# requirements.txt for Python 3.11+
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
nltk>=3.8.0
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-multipart>=0.0.6
joblib>=1.3.0